Refactor Robotron Remix into modular package
