+++ robotron_remix/__init__.py
@@ -1,0 +1,18 @@
+"""Robotron Remix package exports."""
+from .audio import AudioManager
+from .config import *  # noqa: F401,F403
+from .config import __all__ as _config_all
+from .entities import Bullet, Enemy, EnemyAppearance, Player
+from .game import Game
+from .high_scores import HighScoreManager
+
+__all__ = [
+    *_config_all,
+    "AudioManager",
+    "Bullet",
+    "Enemy",
+    "EnemyAppearance",
+    "Game",
+    "HighScoreManager",
+    "Player",
+]
